requirements.txt
fastapi
uvicorn==0.15.0
requests==2.26.0
python-dotenv
pytest-dotenv
httpx